split dmft.utils into submodules
